LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/navigate-intercept-interrupted-expected.txt


FAIL interrupted navigate() promises with intercept() assert_equals: expected 4 but got 2
